refactor: Folder본문 컴포넌트화
